[ISSUE-#129]  - add interface  - delete unused import
